Swap Cluster Data Overview table data
Just load the specific cluster tsv, don't use the cluster_data which
represents ALL of the different tsv files and has a bunch of metadata
related to filtering.

The individual phate_leiden_clustering.tsv for just that
channel_combo/cell_class/leden_resolution is all that we need to show
here.