Add Gene/Cluster search and UX improvements to Cluster page (v3) (#115)
* feat: Sort genes by symbol in cluster analysis

* feat: Style Close Cluster button red

- color red
- float close button to the right
- use english text for the button's action.

* feat: Add border to cluster selection dropdown

* Add gene/cluster search to the top of clusters page.

This works pretty well.
- Pick a gene. The cluster gets set appropriately.
- Pick a cluster. The first gene of the cluster is selected.
- Change side-bar filtering
  - cluster is retained, but might not have any genes
  - gene selection is reset, so it picks the first one.

The gene selection that was already listed under the cluster is still
there and lists the subset of genes for that cluster. If you change that
selection, it mirrors it in the global gene selection. And vice versa
a global gene selection will set the cluster gene selection.

We might be able to try harder to retain the gene selection as sidebar
filters change. But there are no assurances the gene will be there.

* Remove unused methods.

* Move functions. Group data loading.

Comment functional sections.

* Extract cluster_table and feature_table methods from display_cluster

* Add cluster_size_charts

* Add caching for file finding function

* Fix QC Filtering

Pass key to radio button filter for session state

* Sort leiden_resolution numerically ascending.