Add static asset handling for the webserver (#122)
* Add static asset handling for the webserver

Locally using st.download_button is fine. It does load the entire tiff
into memory, but it is only doing them one at a time.

On the webserver it more performant to let it handle the static assets.

So here we define two environment variables, the URL to where we've
mounted the assets in the nginx configuration. And a path to the files
on disk (which we still need to check if they are there, and as a way to
chop up the path appropriately for the url).

* Use static asset links for the tsv downloads as well.